// types/types.d.ts
export interface User extends DocumentNode, basicTable {
  password?: string
  photo?: fileType
  name?: string
  lastname?: string
  email?: string
  privilegeID?: Privilege
  active?: boolean
  tokenExpo?: string
  token?: string
  isActive?: string
  rolName?: string
  canCreateHost: boolean
  allEventWithAuth: boolean
  verifyLogin?: boolean
  lang?: string
  canAccessToApp?: boolean
  canAccessToWeb: ?boolean
  canUseAuthenticator?: boolean
  document: string
  typeDocument: string
  code?: boolean
  phone?: string
  QR?: string
  termporal_Qr?: IWorker_qr_temporal
  group: IGroupWorker[]
  banFinish: string
}
// ...
export type Privilege = {
  _id?: string
  key?: string
  name: privilegeName
  createdAt?: Date
  UpdatedAt?: Date
  permissions?: PermissionsPrivilege[]
}

export type privilegeName = 'Super_admin' | 'admin' | 'host' | 'super_anfitrion'

export type PermissionsPrivilege = {
  sectionID?: Sections['_id']
  read?: boolean
  create?: boolean
  delete?: boolean
  update?: boolean
  canRead?: boolean
  canCreate?: boolean
  canDelete?: boolean
  canUpdate?: boolean
  sectionName?: string
}
// ...
export type fileType = {
  filename: string
  key: string
}

export type verifiedData = {
  photo?: fileType
  documentA?: fileType
  documentB?: fileType
  birthDate?: string
  expirationDate?: string
  sex?: string
  lastName?: string
  firstName?: string
  nationality?: string
  documentNumber?: string
}

export interface IMasterLocation extends Document, basicTable {
  name: string
  address: string
  location: ILocation[] | string[]
  onlyAllowAuthUSers: boolean
}

export interface LocationEntries extends Document, basicTable {
  event: IEvent | string
  contact: IContact | string
  location: ILocation | string
  hourIn: string
}

export interface ILocation extends Document, basicTable {
  state: string
  masterLocation: IMasterLocation | string
  childLocations: String[] | ILocation[]
  parentLocations: String[] | ILocation[]
  address: string
  name: string
  admins: IUser[] | string[]
  operation?: operation
  typeCheck: typeCheck
  device: IHistoryDevice
  host: IUser[] | string[]
  security: IUser[] | string[]
}
export interface IHistoryDevice extends Document, basicTable {
  name: string
  type: typeDevice
  serialNumber: string
  status: statusDevice
  actualLocation: ILocation
  enableVideo: boolean
  enableTalk: boolean
  timeWait: number
  whoDeleted: IUser | string
  deletedDate: string
  origID: string
}
export interface IDevice extends Document, basicTable {
  name: string
  type: typeDevice
  serialNumber: string
  status: statusDevice
  actualLocation: ILocation
  enableVideo: boolean
  enableTalk: boolean
  timeWait: number
}

export type typeLocation = 'classic' | 'touch'

export type graphqlErrors = {
  graphQLErrors: string
  networkError: string
  message: string
  extraInfo: string
}

export interface IGroupWorker extends Document, basicTable {
  name: string
  locations: ILocation[]
  abbreviation: string
  exists: boolean
}
// ...
export interface IWorker_qr_temporal extends Document, basicTable {
  QR: string
  timeEnd: string
  used: boolean
  valid: boolean
  location: ILocation
}